NFT Collection: static frontend
