+++ my-app/dao/styles/Whitelist.module.css
@@ -1,0 +1,16 @@
+.nav {
+    margin: 20px 30px;
+}
+
+.nav img {
+    width: 50px;
+    height: 50px;
+}
+
+.main {
+    margin-top: 50px;
+}
+
+.main img {
+    height: 400px;
+}
